requirements.txt
# Python dependencies for Resume LaTeX Project
PyYAML>=6.0